feat: add device id for mapping controller to allow multiple controllers
Problem:
Currently, device ID isn't store when serializing and deserializing the
controller input actions. As a result, all of the input actions are
only set for device 0, and any other devices are ignored.

Solution:
Store the device ID for the input mapping so input actions for each
controllers are actually persisted.

Risk:
This changes the schema of the input mapping, existing input mappings
saves will need to be invalidated and recreated for the new schema.

Testing Done:
Updated the test with the change and verify they passed.